Extract store into its own module
