+++ src/index.js
@@ -1,31 +1,5 @@
-import PTree, { Path } from "./ptree"
-import MTree, { Model } from "./mtree"
+import Store from "./store"
 
-export { PTree, MTree, Model }
+export default Store
 export connect from "./react/connect"
 export timetravel from "./timetravel"
-
-export default class Store {
-  constructor(initialState = {}, { Engine = PTree } = {}) {
-    this.tree = new Engine({})
-
-    if (initialState.constructor === Promise) {
-      initialState.then(value => this.tree.setRoot(value))
-    } else {
-      this.tree.setRoot(initialState)
-    }
-  }
-
-  subscribe(path, subscriber) {
-    if (arguments.length === 1) {
-      path = Path.root
-      subscriber = arguments[0]
-    }
-
-    return this.tree.subscribe(path, subscriber)
-  }
-
-  get state() {
-    return this.tree.root
-  }
-}
